requirements.txt
termcolor==2.4.0
requests==2.32.3
pyOpenSSL==24.2.1
frida-tools==12.5.1
colorama==0.4.6